+++ PKHeX.Core/Saves/SAV6AO.cs
@@ -154,7 +154,7 @@ namespace PKHeX.Core
         public override void SetDaycareOccupied(int loc, int slot, bool occupied)
         {
             int ofs = loc == 0 ? DaycareOffset : Daycare2;
-            Data[ofs + ((SIZE_STORED + 8) * slot)] = occupied ? 1 : 0;
+            Data[ofs + ((SIZE_STORED + 8) * slot)] = occupied ? (byte)1 : (byte)0;
         }
 
         public override void SetDaycareRNGSeed(int loc, string seed)
@@ -172,7 +172,7 @@ namespace PKHeX.Core
         public override void SetDaycareHasEgg(int loc, bool hasEgg)
         {
             int ofs = loc == 0 ? DaycareOffset : Daycare2;
-            Data[ofs + 0x1E0] = hasEgg ? 1 : 0;
+            Data[ofs + 0x1E0] = hasEgg ? (byte)1 : (byte)0;
         }
 
         public override string JPEGTitle => HasJPPEGData ? string.Empty : StringConverter.GetString6(Data, JPEG, 0x1A);
